Normalizando heranças e banco de dados
