+++ src/ampare-backend/Models/ApplicationDbContext.cs
@@ -11,7 +11,7 @@ namespace ampare_backend.Models
          public DbSet<Cadastro> Cadastros { get; set; }
 
         // DB Ong
-        public DbSet<CadastroOng> CadastroOngs { get; set; }
+        public DbSet<Cadastro> Ongs { get; set; }
 
         //DB Voluntario
         public DbSet<Cadastro> Voluntarios { get; set; }
@@ -25,9 +25,16 @@ namespace ampare_backend.Models
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.Entity<CadastroVoluntario>().ToTable("Voluntarios");
+            modelBuilder.Entity<CadastroOng>().ToTable("Ongs");
         }
 
+        
+
         public DbSet<ampare_backend.Models.CadastroVoluntario> CadastroVoluntario { get; set; }
+
+        
+
+        public DbSet<ampare_backend.Models.CadastroOng> CadastroOng { get; set; }
 
     }
 }
