+++ src/ampare-backend/Views/Shared/_Layout.cshtml
@@ -24,7 +24,7 @@
                         </li>
                         
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="CadastroOngs" asp-action="Index">Ongs</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Ongs" asp-action="Index">Ongs</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Projetos" asp-action="Index">Projetos</a>
